requirements.txt
streamlit==1.31.0
boto3==1.34.51
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
langchain==0.1.9
langchain-aws==0.1.0
duckduckgo-search==5.0.0
plotly==5.18.0
pandas==2.2.0
lxml==5.1.0
